Tests/LibWeb/Text/input/async-scrolling/viewport-wheel-with-nested-scroller.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    body {
        margin: 0;
    }

    #scroller {
        width: 100px;
        height: 100px;
        overflow: scroll;
    }

    #scroller-spacer {
        height: 500px;
    }

    #page-spacer {
        height: 2000px;
    }
</style>
<div id="scroller"><div id="scroller-spacer"></div></div>
<div id="page-spacer"></div>
<script>
    test(() => {
        const state = internals.asyncScrollingState();

        println(`scroll nodes: ${state.scrollNodeCount}`);
        println(`wheel routing admission: ${internals.asyncScrollingStateWheelRoutingAdmission()}`);
        println(`viewport wheel target: ${internals.asyncScrollingStateWheelTargetAt(150, 150, 0, 100)}`);
        println(`scroller wheel target: ${internals.asyncScrollingStateWheelTargetAt(50, 50, 0, 100)}`);
    });
</script>
