Tests/LibWeb/Text/expected/async-scrolling/viewport-wheel-with-nested-scroller.txt

scroll nodes: 2
wheel routing admission: accepted
viewport wheel target: viewport
scroller wheel target: non-viewport
